Implementando aplicação

+++ Produtos.Domain.Core/Intrefaces/Repositories/IRepositoryBase.cs
@@ -9,6 +9,6 @@ namespace Produtos.Domain.Core.Intrefaces.Repositories
         void Update(Tentity obj);
         void Remove(Tentity obj);
         IEnumerable<Tentity> GetAll();
-        Tentity GetById(int id);
+        Tentity GetById(long id);
     }
 }
